core/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
langchain-community
pinecone-client
python-socketio
numpy
